requirements.txt
Pillow
Image
pyautogui
pygetwindow
